Fix logic in whether waveforms are saved

+++ python/src/lardiff/grid_scan.py
@@ -37,6 +37,7 @@ def diffusion_grid_scan(DL_min, DL_max, DL_step, DT_min, DT_max, DT_step,
             test_stat = 0.0
             numvals = 0.0
             all_shifts = np.zeros((num_angle_bins, N_wires))
+
             for k in range(0, num_angle_bins):
                 print('************************ANGLE BIN', k, '***************************')
                 # Generate prediction histograms
@@ -46,14 +47,16 @@ def diffusion_grid_scan(DL_min, DL_max, DL_step, DT_min, DT_max, DT_step,
                     DL, DT, 
                     isdata
                 )
+
                 # Write to pickle file for offline use
                 waveform_file_name = 'output_data/waveforms_{}.pkl'.format('data' if isdata else 'mc')
-                if not os.path.exists(waveform_file_name):
+                if save_waveform_data and not os.path.exists(waveform_file_name):
                     with open(waveform_file_name, 'wb') as fout:
                         pickle.dump((anode_hist, anode_uncert_hist, 
                                      cathode_hist, cathode_uncert_hist, 
                                      pred_hist, pred_uncert_hist), fout)
                         print('Wrote', waveform_file_name)
+
                 # Calculate test statistic for this angle bin
                 temp_test_stat, temp_numvals, shift_vec = calc_test_statistic(
                     anode_hist[k], anode_uncert_hist[k], 
@@ -67,6 +70,7 @@ def diffusion_grid_scan(DL_min, DL_max, DL_step, DT_min, DT_max, DT_step,
                 print('[GRID SCAN] Total test statistic value for angle {}, row {}, col {}, is {}'.format(k, row, col, test_stat))
                 numvals += temp_numvals
                 all_shifts[k, :] = shift_vec
+
                 if verbose:
                     print('test_stat', temp_test_stat)
                     print('    %d %.2f' % (k, temp_test_stat))
